+++ sample/ApplicationInsightsHostingStartupSample/appsettings.json
@@ -15,5 +15,4 @@
       }
     }
   }
-  }
 }
